Refartor movie thumbnail grid to seperate components
